pricealert/requirements.txt
amqp==5.2.0
asgiref==3.6.0
billiard==4.2.0
celery==5.3.6
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
dj-rest-auth==3.0.0
Django==4.1.7
django-celery-beat==2.5.0
django-extensions==3.2.1
django-filter==23.5
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.5
MarkupSafe==2.1.2
packaging==23.0
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0.1
redis==5.0.1
requests==2.28.2
ruamel.yaml==0.17.21
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.13
